# api-doc.yml
openapi: 3.0.0
servers: []
info:
  description: The simple back API of SYP
  version: "1.0.0"
  title: SYP API
  contact:
    email: syp@shareyourproject.fr
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
tags:
  - name: Authentication
    description: For authentication purposes
  - name: Post
  - name: User
  - name: Project
paths:
  /login:
    post:
      tags:
        - Authentication
      summary: Authenticate an existing user
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - password
                - email
      responses:
        '200':
          description: User authenticated
        '422':
          description: Invalid form
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormError'
        '401':
          description: Invalid credentials
  /register:
    post:
      tags:
        - Authentication
      summary: Authenticate and create a new user
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                password_confirmation:
                  type: string
              required:
                - password
                - password_confirmation
                - email
                - first_name
                - last_name
                - username
      responses:
        '200':
          description: User created and authenticated
        '422':
          description: Invalid form
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormError'
  /user:
    get:
      security:
        - cookieAuth: []
      tags:
        - Authentication
      summary: Get the current authenticated user
      responses:
        '200':
          description: User created and authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

  /feed:
    get:
      tags:
        - Post
      summary: Get the post in the user feed
      responses:
        '200':
          description: User created and authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedResponse'
        '401':
          description: Unauthorized


  /posts/{postId}/like:
    put:
      security:
        - cookieAuth: []
      tags:
        - Post
      parameters:
        - in: path
          name: postId
          schema:
            type: integer
          required: true
          description: Numeric ID of the post to like
      summary: Like the current post
      responses:
        '200':
          description: Post has been edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found
        '401':
          description: Unauthenticated

  /posts/{postId}/unlike:
    put:
      security:
        - cookieAuth: []
      tags:
        - Post
      parameters:
        - in: path
          name: postId
          schema:
            type: integer
          required: true
          description: Numeric ID of the post to unlike
      summary: Unlike the current post
      responses:
        '200':
          description: Post has been edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found
        '401':
          description: Unauthenticated

  /projects/search/{query}:
    get:
      tags:
       - Project
      parameters:
        - in: path
          name: query
          schema:
            type: string
          required: true
          description: The search query string
      summary: Search for projects
      responses:
        '200':
          description: List of project
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'

  /users/search/{query}:
    get:
      tags:
       - User
      parameters:
        - in: path
          name: query
          schema:
            type: string
          required: true
          description: The search query string
      summary: Search for user profiles
      responses:
        '200':
          description: List of user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: share_your_project_session
  schemas:
    FormError:
      type: object
      required:
        - message
        - errors
      properties:
        message:
          type: string
        errors:
          type: object
    User:
      type: object
      properties:
        id:
          type: number
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        email_verified_at:
          type: string
        is_admin:
          type: boolean
        bio:
          type: string
        banner_picture:
          type: string
        profile_picture:
          type: string
        title:
          type: string
        updated_at:
          type: string
        created_at:
          type: string

    Tag:
      type: object

    Project:
      type: object
      properties:
        id:
          type: number
        description:
          type: string
        formated_content:
          type: string
        name:
          type: string
        public:
          type: boolean
        owner_id:
          type: number
        technologies:
          type: array
          items:
            $ref: '#/components/schemas/Technologie'
        member_ids:
          type: array
          items:
            type: number
        profile_picture:
          type: string
        banner_picture:
          type: string
        created_at:
          type: string
          format: date
        updated_at:
          type: string
          format: date

    Technologie:
      type: object

    Post:
      type: object
      properties:
        id:
          type: number
        author:
          $ref: '#/components/schemas/User'
        content:
          type: string
        formated_content:
          type: string
        project:
          $ref: '#/components/schemas/Project'
        images_url:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        url:
          type: object
          properties:
            author:
              type: string
            post:
              type: string
        reshared_post:
          type: number
        liked:
          type: boolean
        stats:
          type: object
          properties:
            comments:
              type: number
            like:
              type: number
        updated_at:
          type: string
          format: date
        created_at:
          type: string
          format: date

    FeedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        links:
          type: object
          properties:
            first :
              type: string
            last:
              type: string
            next:
              type: string
            prev:
              type: string
        meta:
          type: object
          properties:
            current_page:
              type: number
            from:
              type: number
            last_page:
              type: number
            links:
              type: array
              items:
                type: object
                properties:
                  active:
                    type: boolean
                  label:
                    type: number
                  url:
                    type: string
            path:
              type: string
            per_page:
              type: number
            to:
              type: number
            total:
              type: number